requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
dnspython==2.4.2
python-whois==0.8.0
shodan==1.30.1
virustotal-python==1.0.1
colorama==0.4.6
tqdm==4.66.1
python-dotenv==1.0.0
flask==3.0.0
google-generativeai==0.3.2
plotly==5.18.0
pandas==2.1.3
flask-wtf==1.2.1
flask-bootstrap4==4.0.2
cryptography==41.0.7
pyOpenSSL==23.3.0 